Voice-synthesis/requirements.txt

tensorflow==1.15.2
numpy==1.13.3
inflect==0.2.5
librosa==0.6.0
scipy==1.0.0
Unidecode==1.0.22
pillow
torch==1.9.0
tensorboardX==1.1
flask==1.1.2
flask-restful
google-cloud-storage
requests